+++ src/wifi/model/extended-capabilities.h
@@ -38,10 +38,10 @@ public:
 
   // Implementations of pure virtual methods of WifiInformationElement
   WifiInformationElementId ElementId () const override;
-  uint8_t GetInformationFieldSize () const override;
+  uint16_t GetInformationFieldSize () const override;
   void SerializeInformationField (Buffer::Iterator start) const override;
-  uint8_t DeserializeInformationField (Buffer::Iterator start,
-                                       uint8_t length) override;
+  uint16_t DeserializeInformationField (Buffer::Iterator start,
+                                        uint16_t length) override;
   /**
    * Set the HT Supported flag.
    *
